# perf: optimize performance of sample_floyd